+++ Lilu/Sources/kern_start.cpp
@@ -268,7 +268,7 @@ bool Configuration::registerPolicy() {
 	return true;
 }
 
-extern "C" kern_return_t kern_start(kmod_info_t *, void *) {
+extern "C" kern_return_t ADDPR(kern_start)(kmod_info_t *, void *) {
 	// Initialise config status
 	atomic_init(&ADDPR(config).initialised, false);
 	// We should be aware of the CPU we run on.
@@ -285,6 +285,6 @@ extern "C" kern_return_t kern_start(kmod_info_t *, void *) {
 	return KERN_SUCCESS;
 }
 
-extern "C" kern_return_t kern_stop(kmod_info_t *, void *) {
+extern "C" kern_return_t ADDPR(kern_stop)(kmod_info_t *, void *) {
 	return ADDPR(config).startSuccess ? KERN_FAILURE : KERN_SUCCESS;
 }