+++ Changelog.md
@@ -3,6 +3,7 @@ Lilu Changelog
 #### v1.3.4
 - Added implicit `eraseCoverageInstPrefix` to `routeMultiple`
 - Fixed user patcher kernel panic when running process via `posix_spawn` without exec
+- Fixed user patcher codesign issues on recent 10.14 versions with SIP
 
 #### v1.3.3
 - Added support for modern AMD device scanning by @AlGreyy