+++ Changelog.md
@@ -1,5 +1,8 @@
 Lilu Changelog
 ==============
+#### v1.3.5
+- Fixed analog audio device detection on certain laptops with Insyde firmware
+
 #### v1.3.4
 - Added implicit `eraseCoverageInstPrefix` to `routeMultiple`
 - Fixed user patcher kernel panic when running process via `posix_spawn` without exec